refactor project - break out route helper funcitons, goal routes, and task routes into separate files. modified create_item helper with additional error checking. updated create_item tests to expect more error details rather than just invalid data. added title filter to get_all_items. confirm all tests still pass.
