+++ app/__init__.py
@@ -30,8 +30,8 @@ def create_app(test_config=None):
     migrate.init_app(app, db)
 
     # Register Blueprints here
-    from .routes import tasks_bp
-    from .routes import goals_bp
+    from app.routes.task import tasks_bp
+    from app.routes.goal import goals_bp
     app.register_blueprint(tasks_bp)
     app.register_blueprint(goals_bp)
